chore(theme): rename tokens.ts to color-tokens.ts
