+++ apps/frontend/src/view/report/pages/table-of-contents.page.tsx
@@ -4,7 +4,7 @@ import { Page } from "#view/report/components/page.report.component.tsx";
 import { Text } from "#view/report/components/text.report.component.tsx";
 import { s1 } from "#view/report/report.style.ts";
 import type { Index, IndexCallback, IndexEntry, ProjectReport } from "#api/types/project.types.ts";
-import { colors } from "#view/wrappers/tokens.ts";
+import { colors } from "#view/wrappers/color-tokens.ts";
 
 interface TableOfContentsPageProps {
     indexCallback: IndexCallback;
